fix uninstall

+++ bhp-cli/bhp-cli/Shell/MainService.cs
@@ -1408,7 +1408,18 @@ namespace Bhp.Shell
                 return true;
             }
             var pluginName = args[1];
-            Directory.Delete(Path.Combine("Plugins", pluginName), true);
+            
+            if (!Plugin.Plugins.Any(u => u.Name == pluginName))
+            {
+                Console.WriteLine("Plugin not found");
+                return true;
+            }
+
+            if (Directory.Exists(Path.Combine("Plugins", pluginName)))
+            {
+                Directory.Delete(Path.Combine("Plugins", pluginName), true);
+            }
+            
             File.Delete(Path.Combine("Plugins", $"{pluginName}.dll"));
             Console.WriteLine($"Uninstall successful, please restart bhp-cli.");
             return true;
